api/requirements.txt
asgiref==3.4.1
Django==3.2
djangorestframework==3.13.1
django-unixtimestampfield==0.5.0
ipdb==0.13.9
langcodes==3.3.0
psycopg2-binary==2.9.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2

